Add storing metrics client in context (#86)
Stealing pattern form our logging library

I'm still trying to figure out how to do the callback to ensure
goroutines are waited on.

For updater it made sense to just do everything in `realMain`. But if we
want to write metrics somewhere within the program (for example maybe we
want to write a metric once we've parsed the command being run).

Ideas:
1. Have the client own a list of callbacks. `WriteMetricAsync()` doesn't
return anything, but there will be a `Close() []error` which will wait
on the goroutines. `WriteMetricAsync()` would append its closer function
to some internal field in the client.

```go
// MetricWriter is a client for reporting metrics about an application's usage.
type MetricWriter interface {
	WriteMetric(ctx context.Context, name string, count int64) error
	WriteMetricAsync(ctx context.Context, name string, count int64)
	Close() []error // only return non-nil errors, maybe this belongs in the struct but not interface?
}
```

Maybe `Close() error ` with a multierr would be better, as it would be
more in line with other libraries implementing `io.Closer`.

2. [My least favorite now that I write it out] Store the functions in
the context using a helper function, much like we do the client. Defer a
function that handles the stored functions in main.
```go
func CleanupFromContext(ctx context.Context) []func() error

// In main
client := metrics.New()
defer func(){
  funcs := CleanupFromContext(ctx)
  // iterate through funcs, handling errors
}
```

3. Just leave as is, every call stack will not return until the metric
is written or timed out. This works well if we only record metrics early
in the call stack, but poorly if its an early branch in the execution
tree.

---------

Signed-off-by: Parker DeWilde <parkerdewilde@gmail.com>
Co-authored-by: Seth Vargo <seth@sethvargo.com>