Simplify client creation and config loading (#112)
This PR does a few things. I can try and break it up if that's easier,
but many of the changes are tightly coupled. The goal of this PR is to
actually loosen those couplings.

1. Remove the construction of installID files from `InstallInfo` and put
it on the client instead. `InstallInfo` really shouldn't have known
about constructing filepaths. The pattern of passing in an "override
file" was a smell.

1. Remove the embedded MetricsConfig struct in the Client (just copy the
required `serverURL` field).

1. Move MetricsConfig validation onto itself (it owns validation, not
the Client).

1. Remove the interim "options" option and use the `Client` directly.
This removes many layers of indirection. Storing things like the
lookuper and nowFunc directly on the `Client` also dramatically simplify
initialization and reduces the number of "things" we need to pass
around. Instead of "installFiles" being an override, they are just a
normal part of the API (with the default value being the default path
that was previously constructed in `InstallInfo`.

1. Use `time.Format(time.RFC3339Nano)` instead of `MarshalText`; yes
`MarshalText` handles an out-of-range date, but that edge case isn't
worth complicating the code. We could use `time.RFC3339` (or any other
representation instead), but `RFC3339Nano` is what `MarhsalText` uses
internally.